carbon/src/androidTest/kotlin/carbon/compose/dropdown/DropdownTest.kt
package carbon.compose.dropdown

import androidx.compose.foundation.layout.Column
import androidx.compose.foundation.layout.fillMaxWidth
import androidx.compose.runtime.SideEffect
import androidx.compose.runtime.mutableStateOf
import androidx.compose.runtime.remember
import androidx.compose.ui.Modifier
import androidx.compose.ui.focus.FocusRequester
import androidx.compose.ui.focus.focusRequester
import androidx.compose.ui.semantics.SemanticsActions
import androidx.compose.ui.test.assert
import androidx.compose.ui.test.assertAll
import androidx.compose.ui.test.assertCountEquals
import androidx.compose.ui.test.assertHeightIsEqualTo
import androidx.compose.ui.test.assertIsDisplayed
import androidx.compose.ui.test.assertIsFocused
import androidx.compose.ui.test.assertIsNotDisplayed
import androidx.compose.ui.test.hasTestTag
import androidx.compose.ui.test.hasText
import androidx.compose.ui.test.isFocusable
import androidx.compose.ui.test.junit4.createComposeRule
import androidx.compose.ui.test.onAllNodesWithTag
import androidx.compose.ui.test.onFirst
import androidx.compose.ui.test.onNodeWithTag
import androidx.compose.ui.test.performClick
import androidx.compose.ui.test.performScrollToNode
import androidx.compose.ui.test.requestFocus
import androidx.compose.ui.unit.dp
import carbon.compose.toList
import org.junit.Assert.assertNotNull
import org.junit.Rule
import org.junit.Test

class DropdownTest {

    // Test:
    // Dropdown is displayed
    // Dropdown is not expanded
    // Dropdown is expanded and options are displayed
    // Chevron icon is displayed
    // Chevron icon is rotated when dropdown is expanded
    // Dropdown options are selectable
    // Placeholder text is displayed
    // Placeholder text changes when an option is selected


    @get:Rule
    val composeTestRule = createComposeRule()

    private val options = (0..9).associateWith { "Option $it" }
    private val minVisibleItems = 4

    @Test
    fun dropdown_field_validateLayout() {
        val isExpanded = mutableStateOf(false)
        composeTestRule.setContent {
            Column(Modifier.fillMaxWidth()) {
                Dropdown(
                    expanded = isExpanded.value,
                    fieldPlaceholderText = "Dropdown",
                    selectedOption = null,
                    options = options,
                    onOptionSelected = {},
                    onExpandedChange = { isExpanded.value = it },
                    onDismissRequest = { isExpanded.value = false },
                    minVisibleItems = minVisibleItems
                )
            }
        }

        composeTestRule.run {
            onNodeWithTag(DropdownTestTags.FIELD)
                .assertIsDisplayed()
                .assertHeightIsEqualTo(40.dp)

            onNodeWithTag(DropdownTestTags.POPUP_CONTENT)
                .assertIsNotDisplayed()

            onNodeWithTag(DropdownTestTags.FIELD_CHEVRON, useUnmergedTree = true)
                .assertIsDisplayed()
        }
    }

    @Test
    fun dropdown_field_validateFocusAbility() {
        val isExpanded = mutableStateOf(false)
        composeTestRule.setContent {
            Column(Modifier.fillMaxWidth()) {
                Dropdown(
                    expanded = isExpanded.value,
                    fieldPlaceholderText = "Dropdown",
                    selectedOption = null,
                    options = options,
                    onOptionSelected = {},
                    onExpandedChange = { isExpanded.value = it },
                    onDismissRequest = { isExpanded.value = false },
                    minVisibleItems = minVisibleItems
                )
            }
        }

        composeTestRule.run {
            onNodeWithTag(DropdownTestTags.FIELD)
                .requestFocus()
                .assertIsFocused()
        }
    }

    @Test
    fun dropdown_optionsPopup_validateLayout() {
        val isExpanded = mutableStateOf(false)
        composeTestRule.setContent {
            Column(Modifier.fillMaxWidth()) {
                Dropdown(
                    expanded = isExpanded.value,
                    fieldPlaceholderText = "Dropdown",
                    selectedOption = null,
                    options = options,
                    onOptionSelected = {},
                    onExpandedChange = { isExpanded.value = it },
                    onDismissRequest = { isExpanded.value = false },
                    minVisibleItems = minVisibleItems
                )
            }
        }

        isExpanded.value = true

        composeTestRule.run {
            onNodeWithTag(DropdownTestTags.POPUP_CONTENT)
                .assertIsDisplayed()

            onAllNodesWithTag(DropdownTestTags.MENU_OPTION).run {
                assertCountEquals(minVisibleItems + 1)
                toList().forEach { it.assertIsDisplayed() }
            }

            onNodeWithTag(DropdownTestTags.POPUP_CONTENT)
                .performScrollToNode(hasText("Option 9"))

            onNode(hasTestTag(DropdownTestTags.MENU_OPTION).and(hasText("Option 9")))
                .assertIsDisplayed()
        }
    }

    @Test
    fun dropdown_optionsPopup_expandAndCollapse() {
        val isExpanded = mutableStateOf(false)
        composeTestRule.setContent {
            Column(Modifier.fillMaxWidth()) {
                Dropdown(
                    expanded = isExpanded.value,
                    fieldPlaceholderText = "Dropdown",
                    selectedOption = null,
                    options = options,
                    onOptionSelected = {},
                    onExpandedChange = { isExpanded.value = it },
                    onDismissRequest = { isExpanded.value = false },
                    minVisibleItems = minVisibleItems
                )
            }
        }

        composeTestRule.run {
            onNodeWithTag(DropdownTestTags.POPUP_CONTENT)
                .assertIsNotDisplayed()

            onNodeWithTag(DropdownTestTags.FIELD)
                .performClick()

            onNodeWithTag(DropdownTestTags.POPUP_CONTENT)
                .assertIsDisplayed()

            // TODO Tried to shrink the dropdown by invoking a touch event on the field, but for
            //  some reason it didn't work.
//            onNodeWithTag(DropdownTestTags.FIELD)
//                .performClick()
            isExpanded.value = false

            onNodeWithTag(DropdownTestTags.POPUP_CONTENT)
                .assertIsNotDisplayed()
        }
    }

    @Test
    fun dropdown_optionsPopup_noOptionSelected_validatePlaceholderDefaultLogic() {
        val isExpanded = mutableStateOf(false)
        val selectedOptionKey = mutableStateOf<Int?>(null)
        val placeholder = "Dropdown"
        composeTestRule.setContent {
            Column(Modifier.fillMaxWidth()) {
                Dropdown(
                    expanded = isExpanded.value,
                    fieldPlaceholderText = placeholder,
                    selectedOption = selectedOptionKey.value,
                    options = options,
                    onOptionSelected = { selectedOptionKey.value = it },
                    onExpandedChange = { isExpanded.value = it },
                    onDismissRequest = { isExpanded.value = false },
                    minVisibleItems = minVisibleItems
                )
            }
        }

        composeTestRule.run {
            onNodeWithTag(DropdownTestTags.FIELD)
                .assert(hasText(placeholder))
                .performClick()

            onAllNodesWithTag(DropdownTestTags.MENU_OPTION)
                .onFirst()
                .performClick()

            onNodeWithTag(DropdownTestTags.FIELD)
                .assert(hasText("Option 0"))
        }
    }

    @Test
    fun dropdown_optionsPopup_optionSelected_validatePlaceholderDefaultLogic() {
        val isExpanded = mutableStateOf(false)
        val selectedOptionKey = mutableStateOf<Int?>(0)
        val placeholder = "Dropdown"
        composeTestRule.setContent {
            Column(Modifier.fillMaxWidth()) {
                Dropdown(
                    expanded = isExpanded.value,
                    fieldPlaceholderText = placeholder,
                    selectedOption = selectedOptionKey.value,
                    options = options,
                    onOptionSelected = { selectedOptionKey.value = it },
                    onExpandedChange = { isExpanded.value = it },
                    onDismissRequest = { isExpanded.value = false },
                    minVisibleItems = minVisibleItems
                )
            }
        }

        composeTestRule.run {
            onNodeWithTag(DropdownTestTags.FIELD)
                .assert(hasText("Option 0"))
        }
    }

    @Test
    fun dropdown_optionsPopup_validateFocusAbility() {
        val selectedOptionKey = mutableStateOf<Int?>(null)

        composeTestRule.setContent {
            val focusRequester = remember { FocusRequester() }

            SideEffect {
                focusRequester.requestFocus()
            }

            Column(
                Modifier
                    .fillMaxWidth()
            ) {
                DropdownContent(
                    options = options,
                    selectedOption = selectedOptionKey.value,
                    colors = DropdownColors.colors(),
                    onOptionSelected = { selectedOptionKey.value = it },
                    modifier = Modifier.weight(1f).focusRequester(focusRequester)
                )
            }
        }

        composeTestRule.run {
            onAllNodesWithTag(DropdownTestTags.MENU_OPTION)
                .assertAll(isFocusable())

            // TODO The first option is focused when the composition completes (with a SideEffect).
            //  However, in this test the focus request seems to be ignored. Investigate why this is
            //  happening.
//            onAllNodesWithTag(DropdownTestTags.MENU_OPTION)
//                .onFirst()
//                .assertIsFocused()

            assertNotNull(
                onAllNodesWithTag(DropdownTestTags.MENU_OPTION)
                    .onFirst()
                    .fetchSemanticsNode()
                    .config
                    .getOrElseNullable(SemanticsActions.RequestFocus) { null }
            )
        }
    }
}
